+++ VideoWeb/VideoWeb.UnitTests/VideoWeb.UnitTests.csproj
@@ -21,7 +21,7 @@
     <PackageReference Include="NUnit" Version="4.1.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.5.0" />
     <PackageReference Include="NBuilder" Version="6.1.0" />
-    <PackageReference Include="VideoApi.Client" Version="3.2.10-pr-0729-0009" />
+    <PackageReference Include="VideoApi.Client" Version="3.2.11-pr-0729-0013" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\VideoWeb.Common\VideoWeb.Common.csproj" />
