+++ VideoWeb/VideoWeb/VideoWeb.csproj
@@ -48,7 +48,7 @@
     </PackageReference>
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.1" />
     <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="6.6.1" />
-    <PackageReference Include="VideoApi.Client" Version="3.2.10-pr-0729-0009" />
+    <PackageReference Include="VideoApi.Client" Version="3.2.11-pr-0729-0013" />
     <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.5" />
     <PackageReference Include="Microsoft.AspNetCore.Http.Connections.Common" Version="8.0.5" />
     <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="8.0.5" />
